+++ modules/School Admin/spaceSettings.php
@@ -19,7 +19,7 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 @session_start();
 
-use Library\Forms\Form;
+use Gibbon\Forms\Form;
 
 if (isActionAccessible($guid, $connection2, '/modules/School Admin/spaceSettings.php') == false) {
     //Acess denied
