+++ src/Forms/FormFactory.php
@@ -17,9 +17,9 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-namespace Library\Forms;
+namespace Gibbon\Forms;
 
-use Library\Forms\FormFactoryInterface;
+use Gibbon\Forms\FormFactoryInterface;
 
 /**
  * FormFactory
