+++ src/Forms/FormRenderer.php
@@ -17,10 +17,10 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-namespace Library\Forms;
+namespace Gibbon\Forms;
 
-use Library\Forms\Form;
-use Library\Forms\FormRendererInterface;
+use Gibbon\Forms\Form;
+use Gibbon\Forms\FormRendererInterface;
 
 /**
  * FormRenderer
