+++ src/Forms/Traits/InputAttributesTrait.php
@@ -17,7 +17,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-namespace Library\Forms\Traits;
+namespace Gibbon\Forms\Traits;
 
 /**
  * Basic HTML Attributes (id, class)
